It now all works.
TODO: Update readme
The program can now generate the intensity values after the convolution,
normalize them to a range between 0 and 255, and test each pixel againts
the high and low threshold.

A pgm is generated for each image with the suffix decribing what stage
of the process the image shows.